add pagination and owner in delNotice controller

+++ controllers/notices/deleteNotices.js
@@ -3,11 +3,18 @@ const Notice = require("../../models/noticesSchema");
 
 const deleteNotices = async (req, res) => {
   const { id } = req.params;
-  const notices = await Notice.findByIdAndRemove(id);
+  const { _id } = req.user;
+  const notice = await Notice.findById(id);
 
-  if (!notices) {
-    throw HttpError(404, "Not found");
+  console.log(notice.owner)
+  console.log(_id)
+
+  if (notice.owner.toString() === _id.toString()) {
+    await Notice.findByIdAndRemove(id);
+  } else {
+    throw HttpError(401, "You do not have access to delete this notice");
   }
+
   res.json({
     message: "Notice deleted",
   });
